Added Deployment builds of jmod.delta, jmod.delta2 and jmod.velocity to build_mac.
I have not touched the built versions of the rest of the externals.

WARNING: jmod.velocity is currently crashing Max. There is something wrong in the new function, but I haven't been able to figure out exactly what yet. Sorry about that. I still wanted to commit, in case Tim might have the time to take a look at it.

%-/

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@956 464221e2-0a0f-0410-983b-f8d57d210c17

+++ Jamoma/library/externals/jmod_velocity/jmod.velocity.cpp
@@ -84,7 +84,7 @@ void *velocity_new(Symbol *s, short ac, Atom *arg)
 		if (ac)									// initialize
 		{
 			if (arg->a_type==A_LONG)
-				x->xn = (double)arg->a_w.w_long;
+				x->xn = (float)arg->a_w.w_long;
 			else if (arg->a_type==A_FLOAT) 
 				x->xn = arg->a_w.w_float;
 			else
@@ -95,7 +95,7 @@ void *velocity_new(Symbol *s, short ac, Atom *arg)
 		}
 		x->lasttime = gettime();
 	}
-	return (x);									// return pointer to instance
+	return(x);									// return pointer to instance
 }
